Achieved control of multiple servos Modified Hand_Control/Hand_Control.ino to implement the simpler serial read method and control the five servos. Removed servoTest() from serial_communication.py as time.sleep() would cause errors with the serial buffer. removed servoTest() from hand_control.py as the method is no longer in serial_communication.py.
Removed drawHandedness() from hands.py as it has already been replaced. Will be added back if the image isn't colorful enough.
Flipped the assignment of finger positions in fingersUp() in hands.py. This number is currently arbitrary and needs to be tested with the rest of the model to determine a proper value.

Added descriptions to the methods in all classes.
Literally changed three characters in main.py.

Signed-off-by: Alex Pereira <alex.pereira.6464@gmail.com>